remove cached property from remaining code
