+++ bim2sim/elements/hvac_elements.py
@@ -10,7 +10,6 @@ from typing import Set, List, Tuple, Generator, Union, Type
 import numpy as np
 
 from bim2sim.kernel.decision import ListDecision, DecisionBunch
-from bim2sim.kernel.decorators import cached_property
 from bim2sim.elements.mapping import condition, attribute
 from bim2sim.elements.base_elements import Port, ProductBased, IFCBased
 from bim2sim.elements.mapping.ifc2python import get_ports as ifc2py_get_ports
@@ -204,19 +203,10 @@ class HVACProduct(ProductBased):
         self.inner_connections: List[Tuple[HVACPort, HVACPort]] \
             = self.get_inner_connections()
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         raise NotImplementedError(f"Please define the expected number of ports "
                                   f"for the class {self.__class__.__name__} ")
-
-    @cached_property
-    def volume(self):
-        if hasattr(self, "net_volume"):
-            if self.net_volume:
-                vol = self.net_volume
-                return vol
-        vol = self.calc_volume_from_ifc_shape()
-        return vol
 
     def get_ports(self) -> list:
         """Returns a list of ports of this product."""
@@ -370,7 +360,7 @@ class HeatPump(HVACProduct):
         unit=ureg.dimensionless
     )
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 4
 
@@ -439,7 +429,7 @@ class Chiller(HVACProduct):
         unit=ureg.kilowatt,
     )
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 4
 
@@ -476,7 +466,7 @@ class CoolingTower(HVACProduct):
         unit=ureg.dimensionless,
     )
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 2
 
@@ -506,7 +496,7 @@ class HeatExchanger(HVACProduct):
         unit=ureg.dimensionless,
     )
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 4
 
@@ -521,7 +511,7 @@ class Boiler(HVACProduct):
         re.compile('Boiler', flags=re.IGNORECASE),
     ]
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 2
 
@@ -686,7 +676,7 @@ class Pipe(HVACProduct):
              'SPOOL']
     }
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 2
 
@@ -794,7 +784,7 @@ class PipeFitting(HVACProduct):
         re.compile('Bend', flags=re.IGNORECASE),
     ]
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return (2, 3)
 
@@ -865,7 +855,7 @@ class Junction(PipeFitting):
         re.compile('Kreuzst(ü|ue)ck', flags=re.IGNORECASE)
     ]
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 3
 
@@ -883,7 +873,7 @@ class SpaceHeater(HVACProduct):
         re.compile('Space.?heater', flags=re.IGNORECASE)
     ]
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 2
 
@@ -979,7 +969,7 @@ class ExpansionTank(HVACProduct):
         re.compile('Ausdehnungs.?gef(ä|ae)(ss|ß)', flags=re.IGNORECASE),
     ]
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 1
 
@@ -1002,7 +992,7 @@ class Storage(HVACProduct):
                                  math.inf * ureg.liter)
     ]
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return float('inf')
 
@@ -1055,7 +1045,7 @@ class Distributor(HVACProduct):
     }
     # TODO why is pipefitting for DH found as Pipefitting and not distributor
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return (2, float('inf'))
 
@@ -1088,7 +1078,7 @@ class Pump(HVACProduct):
              'VERTICALTURBINE']
     }
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 2
 
@@ -1168,7 +1158,7 @@ class Valve(HVACProduct):
              'STOPCOCK']
     }
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 2
 
@@ -1235,7 +1225,7 @@ class ThreeWayValve(Valve):
         re.compile('3-Wege.*?ventil', flags=re.IGNORECASE)
     ]
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 3
 
@@ -1301,7 +1291,7 @@ class Medium(HVACProduct):
         re.compile('Medium', flags=re.IGNORECASE)
     ]
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 0
 
@@ -1309,7 +1299,7 @@ class Medium(HVACProduct):
 class CHP(HVACProduct):
     ifc_types = {'IfcElectricGenerator': ['CHP']}
 
-    @cached_property
+    @property
     def expected_hvac_ports(self):
         return 2
